More world loading work.

+++ Engine/Public/Chunk.h
@@ -8,7 +8,7 @@
 
 class Chunk;
 #include "Actor.h"
-
+#include "WorldModule.h"
 /// <summary> Width of the chunk.</summary>
 static const GLuint CHUNK_WIDTH = 50;
 
@@ -19,7 +19,6 @@ public:
 
 	// freind the actor class so we can share private methods
 	friend Actor;
-
 	/// <summary> Initialises the chunks.</summary>
 	///
 	/// <param name="programIn">    The program in.</param>
@@ -88,7 +87,7 @@ protected:
 	/// <summary> Constructor.</summary>
 	///
 	/// <param name="locationIn"> The location in.</param>
-	Chunk(vec2 locationIn = vec2(0.f, 0.f));
+	Chunk(ivec2 locationIn = ivec2(0, 0));
 
 	/// <summary> The tex uniform handle chunk.</summary>
 	static GLint texUniformHandleChunk;
